+++ plugins/woocommerce-admin/client/layout/navigation.js
@@ -6,9 +6,12 @@ import {
 	WooNavigationItem,
 	getNewPath,
 	getPersistedQuery,
+	getQueryExcludedScreens,
+	getScreenFromPath,
 } from '@woocommerce/navigation';
 import { Link } from '@woocommerce/components';
 import { __ } from '@wordpress/i18n';
+import { useEffect, useState } from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -16,8 +19,28 @@ import { __ } from '@wordpress/i18n';
 import getReports from '../analytics/report/get-reports';
 import { getPages } from './controller';
 import { isWCAdmin } from '../dashboard/utils';
+import { addHistoryListener } from '../navigation/utils';
 
 const NavigationPlugin = () => {
+	const [ persistedQuery, setPersistedQuery ] = useState(
+		getPersistedQuery()
+	);
+
+	const pathIsExcluded = ( path ) =>
+		getQueryExcludedScreens().includes( getScreenFromPath( path ) );
+
+	// Update the persisted queries when history is updated
+	useEffect( () => {
+		return addHistoryListener( () => {
+			setTimeout( () => {
+				if ( pathIsExcluded() ) {
+					return;
+				}
+				setPersistedQuery( getPersistedQuery() );
+			}, 0 );
+		} );
+	}, [] );
+
 	/**
 	 * If the current page is embedded, stay with the default urls
 	 * provided by Navigation because the router isn't present to
@@ -26,7 +49,9 @@ const NavigationPlugin = () => {
 	if ( ! isWCAdmin( window.location.href ) ) {
 		return null;
 	}
+
 	const reports = getReports().filter( ( item ) => item.navArgs );
+
 	const pages = getPages()
 		.filter( ( page ) => page.navArgs )
 		.map( ( page ) => {
@@ -38,7 +63,7 @@ const NavigationPlugin = () => {
 			}
 			return page;
 		} );
-	const persistedQuery = getPersistedQuery( {} );
+
 	return (
 		<>
 			{ pages.map( ( page ) => (
@@ -48,7 +73,11 @@ const NavigationPlugin = () => {
 				>
 					<Link
 						className="components-button"
-						href={ getNewPath( persistedQuery, page.path, {} ) }
+						href={ getNewPath(
+							pathIsExcluded( page.path ) ? {} : persistedQuery,
+							page.path,
+							{}
+						) }
 						type="wc-admin"
 					>
 						{ page.breadcrumbs[ page.breadcrumbs.length - 1 ] }
@@ -63,7 +92,7 @@ const NavigationPlugin = () => {
 					<Link
 						className="components-button"
 						href={ getNewPath(
-							persistedQuery,
+							pathIsExcluded( item.report ) ? {} : persistedQuery,
 							`/analytics/${ item.report }`,
 							{}
 						) }
@@ -77,4 +106,6 @@ const NavigationPlugin = () => {
 	);
 };
 
-registerPlugin( 'wc-admin-navigation', { render: NavigationPlugin } );
+registerPlugin( 'wc-admin-navigation', {
+	render: NavigationPlugin,
+} );