+++ plugins/woocommerce-admin/packages/navigation/src/index.js
@@ -20,6 +20,8 @@ export { getHistory };
 // Export all filter utilities
 export * from './filters';
 
+const TIME_EXCLUDED_SCREENS_FILTER = 'woocommerce_admin_time_excluded_screens';
+
 /**
  * Get the current path from history.
  *
@@ -44,6 +46,31 @@ export const getPersistedQuery = ( query = navUtils.getQuery() ) => {
 		'type',
 	] );
 	return pick( query, params );
+};
+
+/**
+ * Get array of screens that should ignore persisted queries
+ *
+ * @return {Array} Array containing list of screens
+ */
+export const getQueryExcludedScreens = () =>
+	applyFilters( TIME_EXCLUDED_SCREENS_FILTER, [
+		'stock',
+		'settings',
+		'customers',
+		'homescreen',
+	] );
+
+/**
+ * Retrieve a string 'name' representing the current screen
+ *
+ * @param {Object} path Path to resolve, default to current
+ * @return {string} Screen name
+ */
+export const getScreenFromPath = ( path = getPath() ) => {
+	return path === '/'
+		? 'homescreen'
+		: path.replace( '/analytics', '' ).replace( '/', '' );
 };
 
 /**